api/requirements.txt
# Web Server & API Framework
fastapi
uvicorn
python-multipart

# Modern LangChain Integrations
langchain
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community

# Document Processing & RAG Utilities
pypdf
docx2txt
sentence_transformers
pydantic
python-dotenv
requests